20241020_final_quiz_feature

+++ src/api/modules/Setting/Features/v1/Dimensions/DeleteDimensionHandler.cs
@@ -4,6 +4,7 @@ using FSH.Starter.WebApi.Setting.Domain;
 using MediatR;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using FSH.Framework.Core.Exceptions;
 
 namespace FSH.Starter.WebApi.Setting.Features.v1.Dimensions;
 public sealed class DeleteDimensionHandler(
@@ -16,6 +17,12 @@ public sealed class DeleteDimensionHandler(
         ArgumentNullException.ThrowIfNull(request);
         var item = await repository.GetByIdAsync(request.Id, cancellationToken);
         _ = item ?? throw new DimensionNotFoundException(request.Id);
+
+        if (await repository.AnyAsync(new DimensionByFatherIdSpec(item.Id), cancellationToken)) 
+        {
+            throw new ConflictException("Item cannot be deleted as it's being used.");
+        }
+        
         await repository.DeleteAsync(item, cancellationToken);
         logger.LogInformation("Dimension item with id : {ItemId} deleted", item.Id);
     }
